fix: root cause of 'every epoch starts from same point' + ipynb update
Root cause: gradient_checkpoint was wrapping head forward pass, which
prevents attention loss gradients from flowing back to attention_conv.
torch.utils.checkpoint only preserves input/output for recomputation,
but attn_maps are used externally for div/ord/attn_sup losses. Their
gradients cannot correctly reach attention_conv through checkpoint.

Fixes:
1. Remove gradient_checkpoint from head forward in forward_with_attn.
   This is the PRIMARY fix - attention losses can now actually optimize
   the attention mechanism. (Saves ~2GB VRAM trade-off for correctness)

2. Fix cls_loss CutMix: avoid computing head_loss with original label
   when CutMix is active (wasted computation inside autocast).

3. Promote attention loss computations to fp32 (.float()) to prevent
   fp16 underflow in diversity loss dot products and ordering loss
   centroid calculations.

4. Update baseline.ipynb: add new config params to env info cell,
   add resume mode documentation and toggle instructions.

+++ losses/attention.py
@@ -53,8 +53,8 @@ def attention_diversity_loss(attn_maps):
     loss = t.tensor(0.0, device=attn_maps[0].device)
     for i in range(n):
         for j in range(i + 1, n):
-            ai = attn_maps[i].flatten(2)
-            aj = attn_maps[j].flatten(2)
+            ai = attn_maps[i].flatten(2).float()
+            aj = attn_maps[j].flatten(2).float()
             loss = loss + (ai * aj).sum(dim=2).mean()
     return loss / (n * (n - 1) / 2)
 
@@ -66,8 +66,8 @@ def spatial_ordering_loss(attn_maps, bbox_preds=None, bbox_mask=None):
     loss = t.tensor(0.0, device=attn_maps[0].device)
 
     for i in range(len(attn_maps) - 1):
-        ai = attn_maps[i].flatten(2)
-        aj = attn_maps[i + 1].flatten(2)
+        ai = attn_maps[i].flatten(2).float()
+        aj = attn_maps[i + 1].flatten(2).float()
         H, W = attn_maps[i].shape[2], attn_maps[i].shape[3]
         grid_x = t.arange(W, device=ai.device).float()
         grid_x = grid_x.unsqueeze(0).expand(H, W).reshape(1, 1, -1)
